PIM-2977: Replace discriminator map impl by mapped superclass
